refactor: separate enroll from the api
